the fix
the problem is that cannon is loading the deployment of synthetix with updated settings from the `main` preset.

however, because we are changing the settings, the build is no longer the same. so we have to switch to another preset. luckely, the existing version of the spot market package allow for us to supply an updated versio nof the preset easily.

(obv this should not have to be a problem, but vesrion resolution within contexts has always been a bit of a challenge)